feat: Initial commit
Added .gitignore file and README.md file.
Created a setup.sh script to clone a repository, select a branch, build and run a Docker image.
Implemented functions to install Docker Compose and prompt the user for installation.
Checked if Docker and Docker Compose are installed, and printed the URL to access the application.